fixing bugs + renaming
